Fix handling of ELF file symbols.
This fixes `new_string` to handle the case where the string is already
present in the string table. In that case, it should reuse the existing
string, and not increase the size of the string table.

It also creates a new `special_symbols` vector to hold the null and
file symbols, to replace putting them into the `section_symbols` map,
to prevent collisions between a file symbol and a section symbol with
the same name.

And, it fixes the `file_name_is_same_as_symbol_name_issue_31` test to
check for both the SHN_ABS (STT_FILE) symbol, and the STT_FUNC symbol,
which should both be present, with the same name, in that testcase.

Fixes #31.