Port faerie to the targeting crate.
`targeting` is a new crate that provides a vocabulary for describing targets,
including their architecture and binary file format. The patch here replaces
faerie's `Target` enum and `Object` trait with targeting's `Triple` struct.

This has the following advantages:
 - Artifact::emit and Artifact::write no longer have a type parameter, so
   users of faeiry don't have to add extra abstraction in order to support
   changing formats dynamically (as Cretonne currently does [here]).

 - targeting is rich enough that compilers can use it directly. It supports
   the same set of target names that rustc uses. This eliminates the need to
   translate between different target enums when using faerie.

[here]: https://github.com/cretonne/cretonne/blob/master/lib/faerie/src/container.rs#L7

One downside is that this ends up requring the binary format to be specified
in the `Artifact` constructor, rather than being a parameter to `emit` or
`write` at the end. However, if this turns out to be important in practice,
it wouldn't be difficult to add a way to mutate an `Artifact`'s target.